♻️ Recommand View도 적용
