+++ extension/src/contentScript/common/index.ts
@@ -1,5 +1,5 @@
 import ContentBox from './ContentBox';
-import RecommandBox from './RecommandBox';
+import Recommand from './Recommand';
 import Flex from './Flex';
 
-export { ContentBox, RecommandBox, Flex };
+export { ContentBox, Recommand, Flex };
